Create time_series.m
This code is a time series forecasting model using a neural network. The data is loaded and preprocessed for training and testing. The neural network consists of one hidden layer with LSTM cells and a fully connected output layer. The model is trained using the Adam optimizer with various hyperparameters specified in the training options. The model is then used to predict the next 20 values in the time series, which are plotted alongside the original data.